Collect route handlers under sogs.routes.*
