+++ sogs/routes/onion_request.py
@@ -2,10 +2,8 @@ from flask import request, abort
 import json
 from io import BytesIO
 
-from .web import app
-from . import http
-from . import crypto
-from . import utils
+from ..web import app
+from .. import crypto, http, utils
 
 import traceback
